The Password Generator Web Application is a user-friendly tool designed to create secure and customizable passwords.
This project is built using HTML, CSS, and JavaScript, focusing on security and ease of use. The application dynamically generates passwords based on user-selected preferences and provides a simple interface for quick password creation. It also includes a copy feature to easily copy the generated password.

Technology Stack:
HTML for structure
CSS for styling and animations
JavaScript for logic and functionality